Invert the n parameter; add erode_alternative
1. The n parameter for erosion is inverted to mean what it should be: a
pixel is eroded unless it has >= n contacts in 3D space. Previously it
was (27-n) contacts, which does the same thing but does not really make
sense.

2. erode_alternative is a take on the erosion function which does not
use Andrea's 27-dimensional brush but merely sums the surrounding of
non-zero pixels. The result is the same and I thought it would be
faster, but the performance actually varies dramatically based on the
actual number of non-zero pixels in the image (Andrea's is sometimes
slower, sometimes faster, but much more constant).